Module progpow_cpu::progpow

ProgPoW (Programmatic Proof-of-Work) is the Ethereum network’s proposed new Application-Specific Integrated Circuit (ASIC) resistant Proof-of-Work mining algorithm.

ProgPoW’s aim is to reduce the efficiencies of specialized mining devices known as ASIC chips (and accelerated GPU-based setups), and to maximize the performance of General Purpose Hardware (GPUs) to enable more users to compete for new cryptocurrency awarded by the protocol.

ASIC chips are those that are tailored to efficiently mining cryptocurrency based on a specific hashing algorithm.

GPU mining setups are less specialised are struggle to compete for mining rewards.

It would be a change from Ethereum’s current PoW mining algorithm known as Ethash.

ProgPoW audits have been proposed to analyse the efficiency of a ProgPoW ASICs over GPUs and analysis of the economic impact on the Ethereum protocol.

Functions

generate_cdag

keccak_f800_long

keccak_f800_short

progpow

Type Definitions

CDag